Short -> Int. UShort is experimental, use Int to avoid overflow.

+++ src/main/kotlin/com/vadeen/dns/constants/ResourceClass.kt
@@ -1,10 +1,10 @@
 package com.vadeen.dns.constants
 
-sealed class ResourceClass(cls: Short) : DynamicEnum<Short>(cls) {
+sealed class ResourceClass(cls: Int) : DynamicEnum<Int>(cls) {
 
-    class Unknown(cls: Short) : ResourceClass(cls)
+    class Unknown(cls: Int) : ResourceClass(cls)
 
     companion object {
-        fun of(cls: Short): ResourceClass = Unknown(cls)
+        fun of(cls: Int): ResourceClass = Unknown(cls)
     }
 }
